+++ src/utilities.ts
@@ -1,5 +1,6 @@
 'use strict';
 import * as vscode from "vscode";
+import * as fs from 'fs';
 const player = require('node-wav-player');
 
 export class Utilities {
@@ -49,14 +50,50 @@ export class Utilities {
     }
 
     /**
+     * Get a path for storing data that the extension use while it's running 
+     * @param context vscode.ExtensionContext
+     */
+    public static getStoragePath(context:vscode.ExtensionContext): string{
+        // About 'context.globalStoragePath':
+        // An absolute file path in which the extension can store global state. 
+        // The directory might not exist on disk and creation is up to the extension. 
+        // However, the parent directory is guaranteed to be existent.
+        let storagePath:string  = ( context.globalStoragePath ) ? context.globalStoragePath : "";
+        // console.log(`storagePath: ${storagePath}`);
+        if (storagePath !== "" && !fs.existsSync(storagePath)) {
+            console.log(`First time and create dir: ${storagePath}`);
+            fs.mkdirSync(storagePath);
+        }
+        return storagePath;
+    }
+
+    /**
+     * Unlink File 
+     * @param path 
+     */
+    public static unlinkFile(path:string): void {
+        try {
+            fs.unlinkSync(path);
+        //file removed
+        } catch(error) {
+            console.error(error);
+            throw new Error(`unlinkFile failure: ${error.message}`);
+        }
+    }
+
+    /**
      * Play a specified sound file
      * @param soundfile : Wav format sound file
+     * @param cleanup : whether to cleanup sound file after its play out
      */
-    public static playSound(soundfile: string) {
+    public static playSound(soundfile: string, cleanup: boolean = false) {
         player.play({
             path: soundfile,
         }).then(() => {
             // console.log('The sound file started to be played successfully.');
+            if (cleanup){
+                this.unlinkFile(soundfile);
+            }
         }).catch((error:any) => {
             console.error(error);
             throw new Error(`Sound file play failed: ${error.message}`);