+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "vscode-translator-voice" extension will be documented in this file.
 
+## 0.3.0
+- Change a path for temporary sound data from workspace specific directory to global directory (`vscode.ExtensionContext.globalStoragePath`)
+- Unlink sound file after its playout
+
 ## 0.2.0
 - Properties values in the extension app are updated with new configuration values without app restart - [Issue #1](https://github.com/yokawasa/vscode-translator-voice/issues/1)
 - Fixup bug: Toggled Voice state isn't persisted - [Issue #2](https://github.com/yokawasa/vscode-translator-voice/issues/2)